pyproject.toml
[tool.poetry]
name = "mojito-api"
version = "0.2.0"
description = "Repository for Mojito API service."
authors = ["kyomi <me@kyomi.dev>"]
license = "GPL-3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
psycopg2 = "^2.9.6"
djangorestframework = "^3.14.0"
django-environ = "^0.10.0"
itsdangerous = "^2.1.2"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"
black = "^23.3.0"
isort = "^5.12.0"
commitizen = "^3.5.2"
flake8 = "^6.0.0"
mypy = "^1.4.1"
coverage = "^7.2.7"
django-stubs = {version = "^4.2.3", extras = ["compatible-mypy"]}
djangorestframework-stubs = {version = "^3.14.2", extras = ["compatible-mypy"]}

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
sphinx = "^7.0.1"
sphinx-material = "^0.0.35"

[tool.black]
color = true
line-length = 79

[tool.isort]
profile = "black"
line_length = 79

[tool.commitizen]
version = "0.2.0"
version_files = ["pyproject.toml"]
update_changelog_on_bump = true
changelog_incremental = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
